Incorporate retry mechanism
If a pipeline run fails with error `e`, it should
return Retry(e) if and only if the pipeline still
had a retry count above 0.

This will become handy for scheduling mechanism
which can check whether a pipeline is stil eligible
to be retried or if the pipeline run(s) should be
aborted.